feat: Created rust-like error handling with better OpenGL error handling, working on replacing GL_CALL with gl().call();

+++ shaders/shader.vert
@@ -5,12 +5,12 @@ layout(location = 1) in vec2 texture_coordinate;
 layout(location = 2) in uint texture_layer;
 
 out vec2 v_texture_coordinate;
-out uint v_texture_layer;
+out float v_texture_layer;
 uniform mat4 u_model_view_projection;
 
 void main()
 {
     gl_Position = u_model_view_projection * vec4(position.xy, 0.0, 1.0);
     v_texture_coordinate = texture_coordinate;
-    v_texture_layer = texture_layer;
+    v_texture_layer = float(texture_layer);
 }
